Moved Parameter to a proper file and documented, changed some redundant header, added a default dimension when seting a parameter
